trendshop/requirements.txt
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.0
celery==5.4.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
defusedxml==0.7.1
diff-match-patch==20230430
Django==5.0.6
django-allauth==0.62.1
django-ckeditor==6.7.1
django-core==1.4.1
django-environ==0.11.2
django-extensions==3.2.3
django-faker==0.2
django-filter==24.2
django-import-export==4.0.1
django-jazzmin==3.0.0
django-js-asset==2.2.0
django-user-auth-all==0.1
django-widget-tweaks==1.5.0
fake-factory==9999.9.9
Faker==25.4.0
idna==3.6
import-export==0.3.1
kombu==5.3.7
Markdown==3.5.2
novas==3.1.1.5
numpy==1.26.4
oauthlib==3.2.2
pillow==10.3.0
prompt-toolkit==3.0.43
psycopg2==2.9.9
python-dateutil==2.9.0.post0
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==2.0.0
six==1.16.0
socials==0.2.0
sqlparse==0.5.0
tablib==3.5.0
tzdata==2024.1
urllib3==2.2.0
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.6.0
